+++ R/RcppExports.R
@@ -25,12 +25,24 @@ Release__new_func <- function(resource, amount) {
     .Call(`_simmer_Release__new_func`, resource, amount)
 }
 
+ReleaseAll__new <- function(resource) {
+    .Call(`_simmer_ReleaseAll__new`, resource)
+}
+
+ReleaseAll__new_void <- function() {
+    .Call(`_simmer_ReleaseAll__new_void`)
+}
+
 ReleaseSelected__new <- function(id, amount) {
     .Call(`_simmer_ReleaseSelected__new`, id, amount)
 }
 
 ReleaseSelected__new_func <- function(id, amount) {
     .Call(`_simmer_ReleaseSelected__new_func`, id, amount)
+}
+
+ReleaseSelectedAll__new <- function(id) {
+    .Call(`_simmer_ReleaseSelectedAll__new`, id)
 }
 
 SetCapacity__new <- function(resource, value, mod) {
@@ -363,6 +375,14 @@ get_queue_count_ <- function(sim_, names) {
 
 get_queue_count_selected_ <- function(sim_, id) {
     .Call(`_simmer_get_queue_count_selected_`, sim_, id)
+}
+
+get_seized_ <- function(sim_, names) {
+    .Call(`_simmer_get_seized_`, sim_, names)
+}
+
+get_seized_selected_ <- function(sim_, id) {
+    .Call(`_simmer_get_seized_selected_`, sim_, id)
 }
 
 get_selected_ <- function(sim_, id) {