+++ R/trajectory-class.R
@@ -125,11 +125,19 @@ Trajectory <- R6Class("trajectory",
     },
 
     release = function(resource, amount=1, id=0) {
-      check_args(resource=c("string", "NA"), amount=c("number", "function"), id="number")
+      if (missing(resource))
+        return(private$add_activity(ReleaseAll__new_void()))
+      check_args(
+        resource = c("string", "NA"),
+        amount = c("number", "function", "NA"),
+        id = "number"
+      )
       switch(
-        binarise(is.na(resource), is.function(amount)),
+        binarise(is.na(resource), is.na(amount), is.function(amount)),
         private$add_activity(Release__new(resource, amount)),
         private$add_activity(ReleaseSelected__new(id, amount)),
+        private$add_activity(ReleaseAll__new(resource)),
+        private$add_activity(ReleaseSelectedAll__new(id)),
         private$add_activity(Release__new_func(resource, amount)),
         private$add_activity(ReleaseSelected__new_func(id, amount))
       )