+++ src/activity.cpp
@@ -61,6 +61,16 @@ SEXP Release__new_func(const std::string& resource, const Function& amount) {
 }
 
 //[[Rcpp::export]]
+SEXP ReleaseAll__new(const std::string& resource) {
+  return XPtr<Release<int> >(new Release<int>(resource));
+}
+
+//[[Rcpp::export]]
+SEXP ReleaseAll__new_void() {
+  return XPtr<Release<int> >(new Release<int>());
+}
+
+//[[Rcpp::export]]
 SEXP ReleaseSelected__new(int id, int amount) {
   return XPtr<Release<int> >(new Release<int>(id, amount));
 }
@@ -68,6 +78,11 @@ SEXP ReleaseSelected__new(int id, int amount) {
 //[[Rcpp::export]]
 SEXP ReleaseSelected__new_func(int id, const Function& amount) {
   return XPtr<Release<Function> >(new Release<Function>(id, amount));
+}
+
+//[[Rcpp::export]]
+SEXP ReleaseSelectedAll__new(int id) {
+  return XPtr<Release<int> >(new Release<int>(id));
 }
 
 //[[Rcpp::export]]
@@ -392,20 +407,17 @@ SEXP Log__new_func(const Function& message, int level) {
 
 //[[Rcpp::export]]
 int activity_get_count_(SEXP activity_) {
-  XPtr<Activity> activity(activity_);
-  return activity->count;
+  return XPtr<Activity>(activity_)->count;
 }
 
 //[[Rcpp::export]]
 void activity_print_(SEXP activity_, int indent, bool verbose) {
-  XPtr<Activity> activity(activity_);
-  return activity->print(indent, verbose);
+  return XPtr<Activity>(activity_)->print(indent, verbose);
 }
 
 //[[Rcpp::export]]
 SEXP activity_get_next_(SEXP activity_) {
-  XPtr<Activity> activity(activity_);
-  Activity* the_next = activity->get_next();
+  Activity* the_next = XPtr<Activity>(activity_)->get_next();
   if (the_next)
     return XPtr<Activity>(the_next, false);
   return R_NilValue;
@@ -413,8 +425,7 @@ SEXP activity_get_next_(SEXP activity_) {
 
 //[[Rcpp::export]]
 SEXP activity_get_prev_(SEXP activity_) {
-  XPtr<Activity> activity(activity_);
-  Activity* the_prev = activity->get_prev();
+  Activity* the_prev = XPtr<Activity>(activity_)->get_prev();
   if (the_prev)
     return XPtr<Activity>(the_prev, false);
   return R_NilValue;
@@ -430,6 +441,5 @@ void activity_chain_(SEXP first_, SEXP second_) {
 
 //[[Rcpp::export]]
 SEXP activity_clone_(SEXP activity_) {
-  XPtr<Activity> activity(activity_);
-  return XPtr<Activity>(activity->clone());
+  return XPtr<Activity>(XPtr<Activity>(activity_)->clone());
 }